+++ src/main/java/fr/xpdustry/nohorny/NoHornyPlugin.java
@@ -1,28 +1,46 @@
 package fr.xpdustry.nohorny;
 
 import arc.Events;
-import arc.util.Log;
 import fr.xpdustry.distributor.Distributor;
 import fr.xpdustry.distributor.plugin.AbstractPlugin;
+import fr.xpdustry.nohorny.internal.NoHornyConfig;
 import fr.xpdustry.nohorny.logic.HornyLogicBuildContext;
 import fr.xpdustry.nohorny.logic.HornyLogicBuildEvent;
 import fr.xpdustry.nohorny.logic.HornyLogicBuildService;
 import fr.xpdustry.nohorny.logic.HornyLogicBuildType;
 import fr.xpdustry.nohorny.logic.impl.GlobalImageBanService;
 import io.leangen.geantyref.TypeToken;
+import mindustry.Vars;
 import mindustry.game.EventType.BlockBuildEndEvent;
+import mindustry.gen.Call;
+import mindustry.net.Packets.KickReason;
+import mindustry.ui.dialogs.AdminsDialog;
+import mindustry.ui.fragments.PlayerListFragment;
 import mindustry.world.blocks.logic.LogicBlock;
+import net.mindustry_ddns.store.FileStore;
 
 @SuppressWarnings("unused")
 public final class NoHornyPlugin extends AbstractPlugin {
 
-  @SuppressWarnings("FutureReturnValueIgnored")
+  @SuppressWarnings("NullAway.Init")
+  private static FileStore<NoHornyConfig> config;
+
+  /**
+   * Returns the config of the plugin.
+   */
+  public static NoHornyConfig config() {
+    return config.get();
+  }
+
+  @SuppressWarnings({"FutureReturnValueIgnored", "MissingCasesInEnumSwitch"})
   @Override
   public void init() {
+    config = getStoredConfig("config", NoHornyConfig.class);
+
     Distributor.getServicePipeline()
       .registerServiceType(
         TypeToken.get(HornyLogicBuildService.class),
-        GlobalImageBanService.getInstance()
+        new GlobalImageBanService(config.get().isDeepSearchEnabled(), config.get().getLogicBuildCacheSize())
       );
 
     Events.on(BlockBuildEndEvent.class, event -> {
@@ -32,21 +50,26 @@ public final class NoHornyPlugin extends AbstractPlugin {
       if (event.tile.build instanceof LogicBlock.LogicBuild building) {
         building.configure(event.config);
 
-        final var context = new HornyLogicBuildContext(building, player);
         final var future = Distributor.getServicePipeline()
-          .pump(context)
+          .pump(new HornyLogicBuildContext(building, player))
           .through(HornyLogicBuildService.class)
           .getResultAsynchronously();
 
         future.whenComplete((type, throwable) -> {
-          if (type != HornyLogicBuildType.NOT_HORNY) {
-            Log.debug("GIB: Hit @ at (@, @)", player, context.x(), context.y());
-            Events.fire(new HornyLogicBuildEvent(type, player));
-          } else {
-            Log.debug("GIB: Miss @ at (@, @)", player, context.x(), context.y());
-          }
+          if (type != HornyLogicBuildType.NOT_HORNY) Events.fire(new HornyLogicBuildEvent(player));
         });
       }
     });
+
+    switch (config.get().getDefaultAction()) {
+      case KICK -> Events.on(HornyLogicBuildEvent.class, event -> {
+        event.player().kick("You have been kicked for building [pink]NSFW[] logic.");
+      });
+
+      case BAN -> Events.on(HornyLogicBuildEvent.class, event -> {
+        Vars.netServer.admins.banPlayer(event.player().uuid());
+        event.player().kick("You have been banned for building [pink]NSFW[] logic.");
+      });
+    }
   }
 }