+++ src/main/java/fr/xpdustry/nohorny/internal/HornyLogicBuildAction.java
@@ -1,0 +1,8 @@
+package fr.xpdustry.nohorny.internal;
+
+/**
+ * Action performed by default when a horny logic build is detected.
+ */
+public enum HornyLogicBuildAction {
+  NONE, KICK, BAN
+}