+++ src/main/java/fr/xpdustry/nohorny/logic/HornyLogicBuildContext.java
@@ -1,12 +1,16 @@
 package fr.xpdustry.nohorny.logic;
 
 import mindustry.gen.Player;
+import mindustry.world.Tile;
 import mindustry.world.blocks.logic.LogicBlock.LogicBuild;
 import org.jetbrains.annotations.NotNull;
 
-public record HornyLogicBuildContext(int x, int y, @NotNull String code, @NotNull Player player) {
+/**
+ * Context for the {@link HornyLogicBuildService}.
+ */
+public record HornyLogicBuildContext(@NotNull Tile tile, @NotNull String code, @NotNull Player player) {
 
   public HornyLogicBuildContext(final @NotNull LogicBuild building, final @NotNull Player player) {
-    this(building.tileX(), building.tileY(), building.code, player);
+    this(building.tile(), building.code, player);
   }
 }