+++ src/main/java/fr/xpdustry/nohorny/logic/HornyLogicBuildEvent.java
@@ -3,9 +3,9 @@ package fr.xpdustry.nohorny.logic;
 import mindustry.gen.Player;
 import org.jetbrains.annotations.NotNull;
 
-public record HornyLogicBuildEvent(@NotNull HornyLogicBuildType type, @NotNull Player player) {
-  public HornyLogicBuildEvent {
-    if (type == HornyLogicBuildType.NOT_HORNY)
-      throw new IllegalArgumentException("Only fire this event when the type is 'horny' or 'nudity'.");
-  }
+/**
+ * Event thrown when a player build a logic build containing NSFW stuff.
+ */
+public record HornyLogicBuildEvent(@NotNull Player player) {
+
 }