+++ src/main/java/fr/xpdustry/nohorny/logic/HornyLogicBuildType.java
@@ -1,5 +1,8 @@
 package fr.xpdustry.nohorny.logic;
 
+/**
+ * Determines whether the {@link mindustry.world.blocks.logic.LogicBlock.LogicBuild} contained NSFW stuff or not.
+ */
 public enum HornyLogicBuildType {
-  NOT_HORNY, NUDITY, HORNY
+  NOT_HORNY, HORNY
 }