+++ src/main/java/fr/xpdustry/nohorny/logic/impl/GlobalImageBanService.java
@@ -1,13 +1,10 @@
 package fr.xpdustry.nohorny.logic.impl;
 
 import arc.util.Log;
-import arc.util.serialization.Jval;
 import fr.xpdustry.nohorny.logic.HornyLogicBuildContext;
 import fr.xpdustry.nohorny.logic.HornyLogicBuildService;
 import fr.xpdustry.nohorny.logic.HornyLogicBuildType;
-import java.io.BufferedReader;
 import java.io.IOException;
-import java.io.InputStreamReader;
 import java.net.HttpURLConnection;
 import java.net.URL;
 import java.net.URLEncoder;
@@ -22,35 +19,37 @@ import java.util.regex.Pattern;
 import org.checkerframework.checker.nullness.qual.NonNull;
 import org.jetbrains.annotations.NotNull;
 
+/**
+ * Default implementation of the {@link HornyLogicBuildService} backed by GIB, a verification server hosted by Chaotic-Neutral.
+ * It is pretty efficient but quite slow due to individual http requests for each code block.
+ */
 public final class GlobalImageBanService implements HornyLogicBuildService {
 
-  private static final GlobalImageBanService INSTANCE = new GlobalImageBanService();
-
-  private static final MessageDigest messageDigest;
   private static final String GIB_API_ENDPOINT = "http://c-n.ddns.net:9999/bmi/check/?b64hash=";
-  private static final int DEFAULT_CACHE_SIZE = 200;
   private static final int DEFAULT_TIMEOUT = 1000;
   private static final Pattern DRAW_FLUSH_PATTERN = Pattern.compile("^drawflush .*$", Pattern.MULTILINE);
-  private static final Map<String, HornyLogicBuildType> CACHE = new LinkedHashMap<>() {
-    @Override
-    protected boolean removeEldestEntry(final @NotNull Entry<String, HornyLogicBuildType> eldest) {
-      return size() > DEFAULT_CACHE_SIZE;
-    }
-  };
-
-  static {
+  // Thread local because it #handle can be called my other threads from the ServicePipeline
+  private static final ThreadLocal<MessageDigest> MESSAGE_DIGEST_INSTANCE = ThreadLocal.withInitial(() -> {
     try {
-      messageDigest = MessageDigest.getInstance("SHA-256");
+      return MessageDigest.getInstance("SHA-256");
     } catch (NoSuchAlgorithmException e) {
       throw new RuntimeException("Can't find SHA-256 algorithm.", e);
     }
-  }
+  });
 
-  private GlobalImageBanService() {
-  }
+  private final boolean deepSearch;
+  private final int cacheSize;
+  private final Map<String, HornyLogicBuildType> cache;
 
-  public static GlobalImageBanService getInstance() {
-    return INSTANCE;
+  public GlobalImageBanService(final boolean deepSearch, final int cacheSize) {
+    this.deepSearch = deepSearch;
+    this.cacheSize = cacheSize;
+    this.cache = new LinkedHashMap<>(cacheSize) {
+      @Override
+      protected boolean removeEldestEntry(final @NotNull Entry<String, HornyLogicBuildType> eldest) {
+        return size() + 1 > GlobalImageBanService.this.cacheSize;
+      }
+    };
   }
 
   @Override
@@ -59,36 +58,38 @@ public final class GlobalImageBanService implements HornyLogicBuildService {
       return HornyLogicBuildType.NOT_HORNY;
     }
 
-    final var codeBlocks = DRAW_FLUSH_PATTERN.split(context.code(), -1);
+    final var codeBlocks = deepSearch
+      ? DRAW_FLUSH_PATTERN.split(context.code(), -1)
+      : new String[] {context.code()};
+
     for (final var codeBlock : codeBlocks) {
-      final var bytes = messageDigest.digest(codeBlock.getBytes(StandardCharsets.UTF_8));
+      final var bytes = MESSAGE_DIGEST_INSTANCE.get().digest(codeBlock.getBytes(StandardCharsets.UTF_8));
       final var hash = Base64.getEncoder().encodeToString(bytes);
-      final var type = CACHE.computeIfAbsent(hash, h -> {
-        try {
-          Log.info("GIB: Querying hash @", h);
-          final var url = new URL(GIB_API_ENDPOINT + URLEncoder.encode(h, StandardCharsets.UTF_8));
-          final var con = (HttpURLConnection) url.openConnection();
-          con.setConnectTimeout(DEFAULT_TIMEOUT);
-          con.setRequestMethod("GET");
-          con.setDoOutput(true);
-          con.setRequestProperty("Content-Type", "application/json");
+      final var type = cache.computeIfAbsent(hash, this::queryGibDatabase);
+      if (type == HornyLogicBuildType.HORNY) {
+        Log.debug("GIB: Hit @ at (@, @)", context.player().name(), context.tile().x, context.tile().y);
+        return type;
+      }
+    }
 
-          if (con.getResponseCode() == 200) {
-            try (
-              final var stream = con.getInputStream();
-              final var reader = new BufferedReader(new InputStreamReader(stream, StandardCharsets.UTF_8))
-            ) {
-              return Jval.read(reader).getBool("nudity", false) ? HornyLogicBuildType.NUDITY : HornyLogicBuildType.HORNY;
-            }
-          }
-        } catch (IOException e) {
-          Log.debug("An unexpected exception happened while querying the GIB API.", e);
-        }
+    Log.debug("GIB: Miss @ at (@, @)", context.player().name(), context.tile().x, context.tile().y);
+    return HornyLogicBuildType.NOT_HORNY;
+  }
 
-        return HornyLogicBuildType.NOT_HORNY;
-      });
+  private @NotNull HornyLogicBuildType queryGibDatabase(final @NotNull String hash) {
+    try {
+      Log.info("GIB: Querying hash @", hash);
 
-      if (type != HornyLogicBuildType.NOT_HORNY) return type;
+      final var url = new URL(GIB_API_ENDPOINT + URLEncoder.encode(hash, StandardCharsets.UTF_8));
+      final var con = (HttpURLConnection) url.openConnection();
+      con.setConnectTimeout(DEFAULT_TIMEOUT);
+      con.setRequestMethod("GET");
+      con.setDoOutput(true);
+      con.setRequestProperty("Content-Type", "application/json");
+
+      if (con.getResponseCode() == 200) return HornyLogicBuildType.HORNY;
+    } catch (IOException e) {
+      Log.err("An unexpected exception happened while querying the GIB API.", e);
     }
 
     return HornyLogicBuildType.NOT_HORNY;